[ios] another attempt - this time splitting logic between ios, ios safari, and everything else earlier to avoid the async call that messes with the window.open
